Refactor project stucture [WIP]
