+++ client/src/app/components/App/_app.scss
