+++ client/src/app/components/Footer/_Footer.scss
