+++ client/src/app/components/Toolbar/Toolbar.jsx
@@ -6,7 +6,7 @@ import HomeIcon from 'assets/images/home-solid.svg';
 import UsersIcon from 'assets/images/users-solid.svg';
 import ListIcon from 'assets/images/clipboard-list-solid.svg';
 import PlusIcon from 'assets/images/plus-solid.svg';
-import UserBar from 'modules/legacy/UserBar';
+import UserBar from 'modules/legacy/UserBar/UserBar';
 import AppLogo from 'common/components/AppLogo';
 
 const Toolbar = () => (
@@ -48,7 +48,9 @@ const Toolbar = () => (
         <img alt="Plus Icon" className="toolbar__icon-plus" src={PlusIcon} />
       </a>
     </div>
-    <AppLogo />
+    <div className="toolbar__app-logo">
+      <AppLogo />
+    </div>
     <div className="toolbar__right">
       <UserBar />
     </div>
