+++ client/src/app/components/Toolbar/_Toolbar.scss
@@ -5,6 +5,7 @@
   background: $toolbar-bg;
   padding: 8px 16px;
   box-shadow: 0 3px 4px rgba($shadow-color, 0.15);
+  position: relative;
 
   &__left {
     display: flex;
@@ -42,5 +43,12 @@
     position: absolute;
     right: 3px;
     bottom: 3px;
+  }
+
+  &__app-logo {
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
   }
 }
