+++ client/src/app/styles/index.scss
@@ -2,8 +2,7 @@
  * Table of content
  * 1. Global
  * 2. Partials
- * 3. Containers
- * 4. Components
+ * 3. Components
  */
 
 // 1. Global
@@ -15,22 +14,18 @@
 @import './mixins';
 @import './utils';
 
-// 3. Containers
-@import '../../modules/legacy/styles/containers/app';
-
-// 4. Components
-@import '../../modules/legacy/styles/components/AppLogo';
-@import '../../modules/legacy/styles/components/Footer';
-@import '../../modules/legacy/styles/components/Authbox';
-@import '../../modules/legacy/styles/components/googleButton';
-@import '../../modules/legacy/styles/components/Header';
-@import '../../modules/legacy/styles/components/MessageBox';
-@import '../../modules/legacy/styles/components/Preloader';
-@import '../../modules/legacy/styles/components/ProductsContainer';
-@import '../../modules/legacy/styles/components/ProductsList';
-@import '../../modules/legacy/styles/components/InputBar';
-@import '../../modules/legacy/styles/components/UserBar';
-@import '../../modules/legacy/styles/components/SortBox';
-@import '../../modules/legacy/styles/components/VotingBox';
-@import '../../modules/legacy/styles/components/FilterBox';
+// 3. Components
+@import '../components/App/app';
+@import '../../common/components/AppLogo/AppLogo';
+@import '../components/Footer/Footer';
+@import '../../modules/auth-box/Authbox';
+@import '../../common/components/MessageBox/MessageBox';
+@import '../../common/components/Preloader/Preloader';
+@import '../../modules/shopping-list/ProductsContainer/ProductsContainer';
+@import '../../modules/shopping-list/ProductsList/ProductsList';
+@import '../../modules/shopping-list/InputBar/InputBar';
+@import '../../modules/legacy/UserBar/UserBar';
+@import '../../common/components/SortBox/SortBox';
+@import '../../modules/shopping-list/VotingBox/VotingBox';
+@import '../../modules/shopping-list/FilterBox/FilterBox';
 @import '../components/Toolbar/Toolbar';
