+++ client/src/common/components/AppLogo/AppLogo.jsx
