+++ client/src/common/components/AppLogo/_AppLogo.scss
