+++ client/src/common/components/AppLogo/index.js
@@ -1,0 +1,3 @@
+import AppLogo from './AppLogo';
+
+export default AppLogo;
