+++ client/src/common/components/MessageBox/_MessageBox.scss
