+++ client/src/common/components/MessageBox/index.js
@@ -1,0 +1,3 @@
+import MessageBox from './MessageBox';
+
+export default MessageBox;
