+++ client/src/common/components/Preloader/_Preloader.scss
