+++ client/src/common/components/Preloader/index.js
@@ -1,0 +1,3 @@
+import Preloader from './Preloader';
+
+export default Preloader;
