+++ client/src/common/components/SortBox/index.js
@@ -1,0 +1,3 @@
+import SortBox from './SortBox';
+
+export default SortBox;
