+++ client/src/modules/auth-box/_Authbox.scss
@@ -36,3 +36,9 @@
     margin-bottom: 16px;
   }
 }
+
+.google-button {
+  &__img {
+    height: 40px;
+  }
+}
