+++ client/src/modules/auth-box/index.js
@@ -1,0 +1,3 @@
+import AuthBox from './AuthBox';
+
+export default AuthBox;
