+++ client/src/modules/legacy/UserBar/_UserBar.scss
