--- client/src/modules/legacy/styles/components/_Header.scss
@@ -1,33 +1,0 @@
-.app-header {
-  padding: 20px 0 20px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  &__link {
-    color: $brand-color;
-    font-size: 14px;
-    line-height: 1.2;
-    font-weight: 400;
-  }
-
-  &__logo {
-    display: flex;
-    align-items: center;
-    height: 42px;
-  }
-
-  &__icon {
-    height: 20px;
-  }
-
-  &__text {
-    font-size: 28px;
-    line-height: 1.2;
-    display: block;
-    font-weight: 700;
-    text-shadow: 1px 1px 2px $text-light;
-    color: $brand-color;
-    margin-right: 8px;
-  }
-}
