--- client/src/modules/legacy/styles/components/_googleButton.scss
@@ -1,5 +1,0 @@
-.google-button {
-  &__img {
-    height: 40px;
-  }
-}
