+++ client/src/modules/shopping-list/FilterBox/FilterBox.jsx
