+++ client/src/modules/shopping-list/FilterBox/_FilterBox.scss
