+++ client/src/modules/shopping-list/FilterBox/index.js
@@ -1,0 +1,3 @@
+import FilterBox from './FilterBox';
+
+export default FilterBox;
