+++ client/src/modules/shopping-list/InputBar/_InputBar.scss
