+++ client/src/modules/shopping-list/InputBar/index.js
