+++ client/src/modules/shopping-list/ProductsContainer/_ProductsContainer.scss
