+++ client/src/modules/shopping-list/ProductsContainer/index.js
@@ -1,0 +1,3 @@
+import ProductsContainer from './ProductsContainer';
+
+export default ProductsContainer;
