+++ client/src/modules/shopping-list/ProductsList/_ProductsList.scss
