+++ client/src/modules/shopping-list/VotingBox/VotingBox.jsx
