+++ client/src/modules/shopping-list/VotingBox/_VotingBox.scss
