+++ client/src/modules/shopping-list/VotingBox/index.js
@@ -1,0 +1,3 @@
+import VotingBox from './VotingBox';
+
+export default VotingBox;
